// src/api/client.ts
export class AdditionalTask implements IAdditionalTask {
    businessLine!: BusinessLine;
    organization!: string;
    taskDescription!: string;
    projectRequestId!: number;

    constructor(data?: IAdditionalTask) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.businessLine = _data["businessLine"];
            this.organization = _data["organization"];
            this.taskDescription = _data["taskDescription"];
            this.projectRequestId = _data["projectRequestId"];
        }
    }

    static fromJS(data: any): AdditionalTask {
        data = typeof data === 'object' ? data : {};
        let result = new AdditionalTask();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["businessLine"] = this.businessLine;
        data["organization"] = this.organization;
        data["taskDescription"] = this.taskDescription;
        data["projectRequestId"] = this.projectRequestId;
        return data;
    }
}

export interface IAdditionalTask {
    businessLine: BusinessLine;
    organization: string;
    taskDescription: string;
    projectRequestId: number;
}

export class StageData implements IStageData {

    constructor(data?: IStageData) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
    }

    static fromJS(data: any): StageData {
        data = typeof data === 'object' ? data : {};
        let result = new StageData();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        return data;
    }
}

export interface IStageData {
}
// ...
export type BusinessLine = "Architecture" | "Land" | "Transportation";
// ...
export type ChargeType = "Regular" | "Promotional";
// ...
export class ProjectRequestBase implements IProjectRequestBase {
    projectRequestId!: number;
    requestDate!: Date;
    projectRequestStage!: ProjectRequestStage;
    projectName!: string;
    organization!: string;
    businessLine!: BusinessLine;
    requesterEmail!: string;
    optionalManualProjectNumber?: string | undefined;
    chargeType!: ChargeType;
    primaryClient!: string;
    vantagepointWbsn?: string | undefined;
    stage!: StageKind;

    constructor(data?: IProjectRequestBase) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.projectRequestId = _data["projectRequestId"];
            this.requestDate = _data["requestDate"] ? new Date(_data["requestDate"].toString()) : undefined as any;
            this.projectRequestStage = _data["projectRequestStage"];
            this.projectName = _data["projectName"];
            this.organization = _data["organization"];
            this.businessLine = _data["businessLine"];
            this.requesterEmail = _data["requesterEmail"];
            this.optionalManualProjectNumber = _data["optionalManualProjectNumber"];
            this.chargeType = _data["chargeType"];
            this.primaryClient = _data["primaryClient"];
            this.vantagepointWbsn = _data["vantagepointWbsn"];
            this.stage = _data["stage"];
        }
    }

    static fromJS(data: any): ProjectRequestBase {
        data = typeof data === 'object' ? data : {};
        let result = new ProjectRequestBase();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["projectRequestId"] = this.projectRequestId;
        data["requestDate"] = this.requestDate ? this.requestDate.toISOString() : undefined as any;
        data["projectRequestStage"] = this.projectRequestStage;
        data["projectName"] = this.projectName;
        data["organization"] = this.organization;
        data["businessLine"] = this.businessLine;
        data["requesterEmail"] = this.requesterEmail;
        data["optionalManualProjectNumber"] = this.optionalManualProjectNumber;
        data["chargeType"] = this.chargeType;
        data["primaryClient"] = this.primaryClient;
        data["vantagepointWbsn"] = this.vantagepointWbsn;
        data["stage"] = this.stage;
        return data;
    }
}

export interface IProjectRequestBase {
    projectRequestId: number;
    requestDate: Date;
    projectRequestStage: ProjectRequestStage;
    projectName: string;
    organization: string;
    businessLine: BusinessLine;
    requesterEmail: string;
    optionalManualProjectNumber?: string | undefined;
    chargeType: ChargeType;
    primaryClient: string;
    vantagepointWbsn?: string | undefined;
    stage: StageKind;
}

export class ProjectRequest extends ProjectRequestBase implements IProjectRequest {
    additionalTasks!: AdditionalTask[];
    stageData!: StageData;

    constructor(data?: IProjectRequest) {
        super(data);
        if (!data) {
            this.additionalTasks = [];
            this.stageData = new StageData();
        }
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            if (Array.isArray(_data["additionalTasks"])) {
                this.additionalTasks = [] as any;
                for (let item of _data["additionalTasks"])
                    this.additionalTasks!.push(AdditionalTask.fromJS(item));
            }
            this.stageData = _data["stageData"] ? StageData.fromJS(_data["stageData"]) : new StageData();
        }
    }

    static override fromJS(data: any): ProjectRequest {
        data = typeof data === 'object' ? data : {};
        let result = new ProjectRequest();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.additionalTasks)) {
            data["additionalTasks"] = [];
            for (let item of this.additionalTasks)
                data["additionalTasks"].push(item ? item.toJSON() : undefined as any);
        }
        data["stageData"] = this.stageData ? this.stageData.toJSON() : undefined as any;
        super.toJSON(data);
        return data;
    }
}

export interface IProjectRequest extends IProjectRequestBase {
    additionalTasks: AdditionalTask[];
    stageData: StageData;
}

export class ProjectRequestDelta implements IProjectRequestDelta {
    projectRequestStage?: ProjectRequestStage;
    projectName?: string | undefined;
    organization?: string | undefined;
    businessLine?: BusinessLine;
    requesterEmail?: string | undefined;
    optionalManualProjectNumber?: string | undefined;
    chargeType?: ChargeType;
    primaryClient?: string | undefined;
    vantagepointWbsn?: string | undefined;

    constructor(data?: IProjectRequestDelta) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.projectRequestStage = _data["projectRequestStage"];
            this.projectName = _data["projectName"];
            this.organization = _data["organization"];
            this.businessLine = _data["businessLine"];
            this.requesterEmail = _data["requesterEmail"];
            this.optionalManualProjectNumber = _data["optionalManualProjectNumber"];
            this.chargeType = _data["chargeType"];
            this.primaryClient = _data["primaryClient"];
            this.vantagepointWbsn = _data["vantagepointWbsn"];
        }
    }

    static fromJS(data: any): ProjectRequestDelta {
        data = typeof data === 'object' ? data : {};
        let result = new ProjectRequestDelta();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["projectRequestStage"] = this.projectRequestStage;
        data["projectName"] = this.projectName;
        data["organization"] = this.organization;
        data["businessLine"] = this.businessLine;
        data["requesterEmail"] = this.requesterEmail;
        data["optionalManualProjectNumber"] = this.optionalManualProjectNumber;
        data["chargeType"] = this.chargeType;
        data["primaryClient"] = this.primaryClient;
        data["vantagepointWbsn"] = this.vantagepointWbsn;
        return data;
    }
}

export interface IProjectRequestDelta {
    projectRequestStage?: ProjectRequestStage;
    projectName?: string | undefined;
    organization?: string | undefined;
    businessLine?: BusinessLine;
    requesterEmail?: string | undefined;
    optionalManualProjectNumber?: string | undefined;
    chargeType?: ChargeType;
    primaryClient?: string | undefined;
    vantagepointWbsn?: string | undefined;
}
// ...
export type ProjectRequestStage = "Pending" | "Approved" | "Changes Requested";
// ...
export type StageKind = "LONG_LEAD" | "SOURCES_SOUGHT" | "SOLICITATION_RFP_RFQ" | "SUBMITTED" | "SHORTLISTED" | "INTERVIEWED" | "WON_AWAITING_CONTRACT" | "WON_CONTRACT_RECIEVED" | "LOST" | "NO_GO" | "CANCELLED";